backend/requirements.txt
Flask>=2.3.0
flask-cors>=3.1.13
arxiv>=1.4.0
requests>=2.31.0
pdfplumber>=0.9.0
google-generativeai>=0.0.6
python-dotenv>=1.0.0
chromadb>=0.4.0
google-api-core>=2.11.0
